pyproject.toml
[project]
name = "language-learning-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fr-core-news-sm>=3.8.0",
    "google-genai>=1.63.0",
    "ipywidgets>=8.1.8",
    "langchain-core>=1.2.11",
    "langchain-google-genai>=4.2.0",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.9",
    "langgraph>=1.0.8",
    "notebook>=7.5.3",
    "pandas>=3.0.0",
    "pip>=26.0.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "spacy>=3.8.11",
    "spacy-transformers>=1.3.9",
    "typing-extensions>=4.15.0",
    "wordfreq>=3.1.1",
]
